move the calculator to a util function
